requirements.txt
aiogram==3.13.1
bip-utils==2.9.3
qrcode==7.4.2
python-dotenv==1.0.1
Pillow==10.4.0
aiohttp==3.9.5
watchdog==4.0.1

# Testing dependencies
pytest==8.3.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
